requirements.txt
fastapi
cassandra-driver
uvicorn